+++ packages/cleanly_architected_core/test/interactor/mutation_test.dart
@@ -1,6 +1,6 @@
 import 'package:cleanly_architected_core/src/data_source/params.dart';
 import 'package:cleanly_architected_core/src/entity/equatable_entity.dart';
-import 'package:cleanly_architected_core/src/interactor/form.dart';
+import 'package:cleanly_architected_core/src/interactor/mutation.dart';
 import 'package:cleanly_architected_core/src/repository/remote_mutation_repository.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
